app/requirements.txt
fastapi[all]
uvicorn
databases
sqlalchemy
python-dotenv
asyncpg